fix(ci): stop the pipeline breaking on a calendar day, an MSVC warning and a missing GIR
Three red jobs on the last push, three causes:

- **C++ lint / gen.** `gen/map-status.php` stamped `date('Y-m-d')` into docs/GTK3-MAP.md on
  every run, so `./ci.sh --only=gen` - which regenerates and fails on any diff - failed on the
  first CI run made on a later day than the file was committed. The date is now kept as the
  file carries it and advances only when the status column or the counts really changed.
- **Windows.** `gsk_container_node_new()` takes a `guint` count and the override passed
  `nodes.size()`: GCC is silent, MSVC /W3 is not, and the warning gate failed the build. Cast
  it, as every other list override already does.
- **release.yml** ran the full `./ci.sh` (gen included) with only gir1.2-gtk-4.0 installed,
  which the WebKit wave turned into a hard failure - the generator needs WebKit-6.0.gir.
  libwebkitgtk-6.0-dev is in that job's package list now.

Both gates also say more when they fail: the gen stage names the files it regenerated and
prints the first lines of the diff (on a runner the tree is gone with the job), and the PHPUnit
runs stream `--log-events-text`, whose tail is printed when the job fails - a segfault kills
PHPUnit before it writes any JUnit file, so nothing named the test until now.

That last one is for the fourth failure, which is *not* fixed here: PHP 8.5 ZTS died ~46% into
the suite on run 34085363767, and the same job re-run on the same commit passed. Recorded in
docs/TODO.md; the event log will name the test the next time it happens.

+++ src/Gsk/GskContainerNode.cpp
@@ -64,7 +64,9 @@ ZEND_METHOD(Gtk4_GskContainerNode, __construct) {
     nodes.push_back(static_cast<GskRenderNode *>(node));
   }
   ZEND_HASH_FOREACH_END();
-  GskRenderNode *self = gsk_container_node_new(nodes.data(), nodes.size());
+  // The count is a guint, not a gsize: MSVC's /W3 fails the Windows build on the implicit
+  // narrowing (GCC does not warn), so it is spelled out here as everywhere else.
+  GskRenderNode *self = gsk_container_node_new(nodes.data(), static_cast<guint>(nodes.size()));
   fundamental_adopt(fundamental_from_zval(ZEND_THIS), GSK_TYPE_CONTAINER_NODE, self);
 }
